feat(core): centralize remote spec parsing
