+++ crates/cli/src/client/run.rs
@@ -57,9 +57,10 @@ pub(crate) fn run_client(opts: ClientOpts, matches: &ArgMatches) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::utils::{RemoteSpec, parse_bool, parse_remote_spec};
+    use crate::utils::parse_bool;
     #[allow(unused_imports)]
     use crate::{EngineError, cli_command, spawn_daemon_session};
+    use crate::{RemoteSpec, parse_remote_spec};
     use clap::{FromArgMatches, Parser};
     #[allow(unused_imports)]
     use daemon::authenticate;
